chg: [parsers] advanced taskinfo parsing fixes #51
